Leonid/requirements.txt
Flask==2.0.3
PyJWT==2.3.0
llama-cpp-python==0.1.49

